+++ packages/fe-scripts/bin/setup-husky.js
@@ -2,7 +2,7 @@
 
 import { dirname, join } from 'path';
 import { fileURLToPath } from 'url';
-import { setupHusky } from '../scripts/setup-husky.js';
+import { setupHusky } from '../dist/es/scripts/setup-husky.js';
 
 async function main() {
   const commitlintPath = join(dirname(fileURLToPath(import.meta.url)), '../');