+++ packages/fe-scripts/CHANGELOG.md
@@ -1,5 +1,24 @@
 
 
+## [0.5.1](https://github.com/qlover/fe-base/compare/fe-scripts-v0.4.4...fe-scripts-v0.5.1) (2024-12-17)
+
+
+### Features
+
+* fe code2md use typescript ([#180](https://github.com/qlover/fe-base/issues/180)) ([5f086b1](https://github.com/qlover/fe-base/commit/5f086b11fbe98aa410996ce95f8c0b5b5f06b6e5))
+
+## [0.4.4](https://github.com/qlover/fe-base/compare/fe-scripts-v0.4.3...fe-scripts-v0.4.4) (2024-12-16)
+
+
+### Features
+
+* fe utils interface ([#172](https://github.com/qlover/fe-base/issues/172)) ([ad32dd8](https://github.com/qlover/fe-base/commit/ad32dd815d80cf839103ae08ee5174ad73522624))
+
+
+### Bug Fixes
+
+* eslint config ([#171](https://github.com/qlover/fe-base/issues/171)) ([fc14f4b](https://github.com/qlover/fe-base/commit/fc14f4b2fb3196394d3b470eaa25bcfb428c7861))
+
 ## [0.4.3](https://github.com/qlover/fe-base/compare/fe-scripts-v0.4.2...fe-scripts-v0.4.3) (2024-12-10)
 
 ## [0.4.2](https://github.com/qlover/fe-base/compare/fe-scripts-v0.4.1...fe-scripts-v0.4.2) (2024-12-04)